backend/requirements.txt
Django
pytz
djangorestframework
markdown
django-filter
django-extra-fields
drf_yasg
pillow
PyYAML